chore: refactor AssetOverview Swap and Bridge buttons to use custom hook for reusability
